+++ src/Text/Show/Text/Data.hs
@@ -1,0 +1,24 @@
+-----------------------------------------------------------------------------
+-- |
+-- Module      :  Text.Show.Text.Data
+-- Copyright   :  (C) 2014 Ryan Scott
+-- License     :  BSD-style (see the file LICENSE)
+-- Maintainer  :  Ryan Scott
+-- Stability   :  Experimental
+-- Portability :  GHC
+-- 
+-- Imports 'Show' functions for @Data@ modules.
+----------------------------------------------------------------------------
+module Text.Show.Text.Data () where
+
+import Text.Show.Text.Data.Array      ()
+import Text.Show.Text.Data.Char       ()
+import Text.Show.Text.Data.Containers ()
+import Text.Show.Text.Data.Either     ()
+import Text.Show.Text.Data.Enum       ()
+import Text.Show.Text.Data.Floating   ()
+import Text.Show.Text.Data.Integral   ()
+import Text.Show.Text.Data.List       ()
+import Text.Show.Text.Data.Maybe      ()
+import Text.Show.Text.Data.Text       ()
+import Text.Show.Text.Data.Tuple      ()