+++ src/Text/Show/Text/Foreign.hs
@@ -1,0 +1,14 @@
+-----------------------------------------------------------------------------
+-- |
+-- Module      :  Text.Show.Text.Foreign
+-- Copyright   :  (C) 2014 Ryan Scott
+-- License     :  BSD-style (see the file LICENSE)
+-- Maintainer  :  Ryan Scott
+-- Stability   :  Experimental
+-- Portability :  GHC
+-- 
+-- Imports 'Show' functions for @Foreign@ modules.
+----------------------------------------------------------------------------
+module Text.Show.Text.Foreign () where
+
+import Text.Show.Text.Foreign.Ptr ()